health-main/requirements.txt
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
tensorflow>=2.15.0
opencv-python>=4.8.0
pillow>=10.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
joblib>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
imbalanced-learn>=0.11.0
nibabel>=5.1.0
pydicom>=2.4.0
scipy>=1.11.0
scikit-image>=0.21.0